fix: sizes and colors
